Deregister AMI within time range
This script will deregister AMI's in a specified time frame and delete the associated Snapshots.

NOTE - Make sure to set the time rage and region as necessary before executing the script. You must also bear in mind that if you have two AMI's sharing a Snapshot and one of the AMI's is outside of the time frame, you will run into issues.

Instructions for execution of the script:
1. Install the AWS CLI
2. Configure the AWS CLI with the appropriate IAM permissions (User/Role)
3. Create a file for the script (I'll call it 'AMI-deregister.sh')
4. Set the region appropriately (it's a variable at the beginning of the script)
5. Set the date time frame as necessary (Line 6 of the script) (year-month-date)
6. Give the script executable permissions ($ chmod u+rx AMI-deregister.sh)
7. Execute the script ($ ./AMI-deregister.sh)
Done!